Get barebones program up and running again Todo: Relative vs absolute imports?
