+++ src/main.py
@@ -1,9 +1,7 @@
-import init
-import load_points
-import network
-import net_modification
-import interactive_analysis
-import plotting
+from init import data_loading
+import objects.load_points as load_points
+import objects.net_modification as net_modification
+from analysis import interactive_analysis
 import utilities
 
 print()
@@ -12,8 +10,8 @@ print("##                              Generic Load Modelling                   
 print("#############################################################################################")
 print()
 
-STR_CONFIG_PATH = "example_data\\example_config.toml"
-dict_config, dict_data, dict_network = init.initialize_config_and_data(
+STR_CONFIG_PATH = "in_data/example_data/example_config.toml"
+dict_config, dict_data, dict_network = data_loading.initialize_config_and_data(
     STR_CONFIG_PATH)
 
 # Network datastructures
