+++ src/modelling/modelling.py
@@ -1,4 +1,4 @@
-from models import toenne
+from modelling.models import toenne
 
 
 def model_load(dict_modelling_config, dict_data_ts):
